moved Interfaces and added delete domain endpoint
